requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
anyio==4.8.0
APScheduler==3.6.3
blinker==1.9.0
cachetools==4.2.2
certifi==2024.8.30
charset-normalizer==3.4.0
chatterbot-corpus==1.2.0
click==8.1.7
cmake==3.31.1
colorama==0.4.6
comtypes==1.4.8
contourpy==1.3.1
cycler==0.12.1
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
fonttools==4.55.2
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.68.1
gTTS==2.5.4
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
priority==2.0.0
protobuf==5.29.1
psutil==6.1.0
py-cpuinfo==9.0.0
PyAudio==0.2.14
pybboxes==0.1.6
pygame==2.6.1
pyparsing==3.2.0
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pyttsx3==2.98
pytz==2024.2
pywin32==308
PyYAML==3.13
Quart==0.20.0
rarfile==4.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.6
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
SpeechRecognition==3.12.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tornado==6.1
tqdm==4.67.1
transformers==4.47.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ultralytics==8.3.44
ultralytics-thop==2.0.12
unrar==0.4
urllib3==2.2.3
waitress==3.0.2
Werkzeug==3.1.3
wsproto==1.2.0
